Switch to f32 for the sample type. Now using the chosen output device rather than the default.

+++ src/libs/audio/audio_devices.rs
@@ -21,7 +21,7 @@ pub fn list_audio_devices(pa: &PortAudio) -> Result<i32, Box<dyn Error>> {
         let input_params = pa::StreamParameters::<i16>::new(idx, in_channels, INTERLEAVED, LATENCY);
         let out_channels = info.max_output_channels;
         let output_params =
-            pa::StreamParameters::<i16>::new(idx, out_channels, INTERLEAVED, LATENCY);
+            pa::StreamParameters::<f32>::new(idx, out_channels, INTERLEAVED, LATENCY);
         let in_48k_supported = pa.is_input_format_supported(input_params, SAMPLE_RATE).is_ok();
         let out_48k_supported = pa.is_output_format_supported(output_params, SAMPLE_RATE).is_ok();
         let support_48k = if (in_channels > 0 && in_48k_supported) || (out_channels > 0 && out_48k_supported) { "48000Hz supported" } else { "48000Hz not supported" };
@@ -37,7 +37,7 @@ pub fn output_audio_device_exists(pa: &PortAudio, dev_name: &str) -> Result<bool
 
         let out_channels = info.max_output_channels;
         let output_params =
-            pa::StreamParameters::<i16>::new(idx, out_channels, INTERLEAVED, LATENCY);
+            pa::StreamParameters::<f32>::new(idx, out_channels, INTERLEAVED, LATENCY);
         let out_48k_supported = pa.is_output_format_supported(output_params, SAMPLE_RATE).is_ok();
         if info.name == dev_name && out_channels > 0 && out_48k_supported {
             return Ok(true)
@@ -52,7 +52,7 @@ pub fn input_audio_device_exists(pa: &PortAudio, dev_name: &str) -> Result<bool,
 
         let in_channels = info.max_input_channels;
         let input_params =
-            pa::StreamParameters::<i16>::new(idx, in_channels, INTERLEAVED, LATENCY);
+            pa::StreamParameters::<f32>::new(idx, in_channels, INTERLEAVED, LATENCY);
         let in_48k_supported = pa.is_input_format_supported(input_params, SAMPLE_RATE).is_ok();
         if info.name == dev_name && in_channels > 0 && in_48k_supported {
             return Ok(true)
@@ -61,13 +61,13 @@ pub fn input_audio_device_exists(pa: &PortAudio, dev_name: &str) -> Result<bool,
     Ok(false)
 }
 
-pub fn open_output_audio_device(pa: &PortAudio, dev_name: &str) -> Result<OutputStreamSettings<i16>, Box<dyn Error>> {
+pub fn open_output_audio_device(pa: &PortAudio, dev_name: &str) -> Result<OutputStreamSettings<f32>, Box<dyn Error>> {
     for device in pa.devices()? {
         let (idx, info) = device?;
 
         let out_channels = info.max_output_channels;
         let output_params =
-            pa::StreamParameters::<i16>::new(idx, out_channels, INTERLEAVED, LATENCY);
+            pa::StreamParameters::<f32>::new(idx, out_channels, INTERLEAVED, LATENCY);
         let out_48k_supported = pa.is_output_format_supported(output_params, SAMPLE_RATE).is_ok();
         if info.name == dev_name && out_channels > 0 && out_48k_supported {
             let settings = OutputStreamSettings::new(output_params, SAMPLE_RATE, FRAMES_PER_BUFFER);
